(#933) Updated tests
Renamed class and method names acording to conventions
Test the full content of the dispatched notifications
Test that notification are NOT dispatched when missing contacts
